feat(bot): added scrapper client
HW3

+++ scrapper/src/main/java/ru/tinkoff/edu/scrapper/dto/LinkResponse.java
@@ -1,4 +1,4 @@
 package ru.tinkoff.edu.scrapper.dto;
 
-public record LinkResponse(int id, String url) {
+public record LinkResponse(Long id, String url) {
 }
